+++ README.md
@@ -9,7 +9,7 @@
     - [ ] 중복된 일시와 테마로 예약할 수 없다.
 - 방탈출 테마 추가
   - [x] 테마 테이블 및 예약 테이블 외래키 추가
-  - [ ] 관리자 기능 구현
+  - 관리자 기능 구현
     - [ ] 전체 조회
     - [ ] 테마 추가
       - [ ] 동일한 이름의 테마를 추가할 수 없다.
