+++ src/main/java/roomescape/presentation/AdminPageController.java
@@ -1,0 +1,28 @@
+package roomescape.presentation;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public final class AdminPageController {
+
+    @GetMapping("")
+    public String home() {
+        return "home/index";
+    }
+
+    @GetMapping("admin")
+    public String admin() {
+        return "admin/index";
+    }
+
+    @GetMapping("admin/reservation")
+    public String reservation() {
+        return "admin/reservation";
+    }
+
+    @GetMapping("admin/time")
+    public String time() {
+        return "admin/time";
+    }
+}
